Enable context menus for all widgets.
This is a straightforward move of code from Fields to Widgets, which is
possible now the _add_event_notifiers, etc. has been standardized,
