+++ pyface/fields/i_field.py
@@ -11,7 +11,7 @@
 """ The field interface. """
 
 
-from traits.api import Any, HasTraits, Instance, Str
+from traits.api import Any, HasTraits
 
 from pyface.i_layout_widget import ILayoutWidget
 
@@ -26,21 +26,12 @@ class IField(ILayoutWidget):
     #: The value held by the field.
     value = Any()
 
-    #: An optional context menu for the field.
-    context_menu = Instance("pyface.action.menu_manager.MenuManager")
-
-    def show_context_menu(self, x, y):
-        """ Create and show the context menu at a position. """
-
 
 class MField(HasTraits):
     """ The field mix-in. """
 
     #: The value held by the field.
     value = Any()
-
-    #: An optional context menu for the field.
-    context_menu = Instance("pyface.action.menu_manager.MenuManager")
 
     # ------------------------------------------------------------------------
     # IWidget interface
@@ -50,11 +41,6 @@ class MField(HasTraits):
         """ Set up toolkit-specific bindings for events """
         super()._add_event_listeners()
         self.observe(self._value_updated, "value", dispatch="ui")
-        self.observe(
-            self._context_menu_updated, "context_menu", dispatch="ui"
-        )
-        if self.control is not None and self.context_menu is not None:
-            self._observe_control_context_menu()
 
     def _remove_event_listeners(self):
         """ Remove toolkit-specific bindings for events """
@@ -114,34 +100,9 @@ class MField(HasTraits):
         """ Toolkit specific method to change the control value observer. """
         raise NotImplementedError()
 
-    def _observe_control_context_menu(self, remove=False):
-        """ Toolkit specific method to change the control menu observer.
-
-        This should use _handle_control_context_menu as the event handler.
-        """
-        raise NotImplementedError()
-
-    def _handle_control_context_menu(self, event):
-        """ Handle a context menu event.
-
-        This should call show_context_menu with appropriate position x and y
-        arguments.
-
-        The function signature will likely vary from toolkit to toolkit.
-        """
-        raise NotImplementedError()
-
     # Trait change handlers -------------------------------------------------
 
     def _value_updated(self, event):
         value = event.new
         if self.control is not None:
             self._set_control_value(value)
-
-    def _context_menu_updated(self, event):
-
-        if self.control is not None:
-            if event.new is None:
-                self._observe_control_context_menu(remove=True)
-            if event.old is None:
-                self._observe_control_context_menu()
