+++ pyface/ui/qt4/fields/field.py
@@ -11,9 +11,8 @@
 """ The Qt-specific implementation of the text field class """
 
 
-from traits.api import Any, Instance, Str, provides
+from traits.api import Any, provides
 
-from pyface.qt.QtCore import Qt
 from pyface.fields.i_field import IField, MField
 from pyface.ui.qt4.layout_widget import LayoutWidget
 
@@ -27,29 +26,3 @@ class Field(MField, LayoutWidget):
 
     #: The value held by the field.
     value = Any()
-
-    #: An optional context menu for the field.
-    context_menu = Instance("pyface.action.menu_manager.MenuManager")
-
-    # ------------------------------------------------------------------------
-    # Private interface
-    # ------------------------------------------------------------------------
-
-    def _observe_control_context_menu(self, remove=False):
-        """ Toolkit specific method to change the control menu observer. """
-        if remove:
-            self.control.setContextMenuPolicy(Qt.DefaultContextMenu)
-            self.control.customContextMenuRequested.disconnect(
-                self._handle_control_context_menu
-            )
-        else:
-            self.control.customContextMenuRequested.connect(
-                self._handle_control_context_menu
-            )
-            self.control.setContextMenuPolicy(Qt.CustomContextMenu)
-
-    def _handle_control_context_menu(self, pos):
-        """ Signal handler for displaying context menu. """
-        if self.control is not None and self.context_menu is not None:
-            menu = self.context_menu.create_menu(self.control)
-            menu.show(pos.x(), pos.y())
