+++ pyface/ui/qt4/widget.py
@@ -11,10 +11,9 @@
 
 
 from pyface.qt import QtCore
-
+from pyface.qt.QtCore import Qt
 
 from traits.api import Any, Bool, HasTraits, Instance, Str, provides
-
 
 from pyface.i_widget import IWidget, MWidget
 
@@ -41,6 +40,9 @@ class Widget(MWidget, HasTraits):
 
     #: A tooltip for the field.
     tooltip = Str()
+
+    #: An optional context menu for the widget.
+    context_menu = Instance("pyface.action.menu_manager.MenuManager")
 
     # Private interface ----------------------------------------------------
 
@@ -122,6 +124,25 @@ class Widget(MWidget, HasTraits):
     def _set_control_tooltip(self, tooltip):
         """ Toolkit specific method to set the control's tooltip. """
         self.control.setToolTip(tooltip)
+
+    def _observe_control_context_menu(self, remove=False):
+        """ Toolkit specific method to change the control menu observer. """
+        if remove:
+            self.control.setContextMenuPolicy(Qt.DefaultContextMenu)
+            self.control.customContextMenuRequested.disconnect(
+                self._handle_control_context_menu
+            )
+        else:
+            self.control.customContextMenuRequested.connect(
+                self._handle_control_context_menu
+            )
+            self.control.setContextMenuPolicy(Qt.CustomContextMenu)
+
+    def _handle_control_context_menu(self, pos):
+        """ Signal handler for displaying context menu. """
+        if self.control is not None and self.context_menu is not None:
+            menu = self.context_menu.create_menu(self.control)
+            menu.show(pos.x(), pos.y())
 
     # Trait change handlers --------------------------------------------------
