+++ pyface/ui/wx/fields/field.py
@@ -11,9 +11,7 @@
 """ The Wx-specific implementation of the text field class """
 
 
-from traits.api import Any, Instance, Str, provides
-
-import wx
+from traits.api import Any, provides
 
 from pyface.fields.i_field import IField, MField
 from pyface.ui.wx.layout_widget import LayoutWidget
@@ -28,25 +26,3 @@ class Field(MField, LayoutWidget):
 
     #: The value held by the field.
     value = Any()
-
-    #: An optional context menu for the field.
-    context_menu = Instance("pyface.action.menu_manager.MenuManager")
-
-    # ------------------------------------------------------------------------
-    # Private interface
-    # ------------------------------------------------------------------------
-
-    def _observe_control_context_menu(self, remove=False):
-        """ Toolkit specific method to change the control menu observer. """
-        if remove:
-            self.control.Unbind(
-                wx.EVT_CONTEXT_MENU, handler=self._handle_context_menu
-            )
-        else:
-            self.control.Bind(wx.EVT_CONTEXT_MENU, self._handle_context_menu)
-
-    def _handle_control_context_menu(self, event):
-        """ Signal handler for displaying context menu. """
-        if self.control is not None and self.context_menu is not None:
-            menu = self.context_menu.create_menu(self.control)
-            self.control.PopupMenu(menu)
