+++ pyface/ui/wx/widget.py
@@ -12,9 +12,9 @@
 """ Enthought pyface package component
 """
 
+import wx
 
-from traits.api import Any, Bool, HasTraits, Str, provides
-
+from traits.api import Any, Bool, HasTraits, Instance, Str, provides
 
 from pyface.i_widget import IWidget, MWidget
 
@@ -41,6 +41,9 @@ class Widget(MWidget, HasTraits):
 
     #: A tooltip for the widget.
     tooltip = Str()
+
+    #: An optional context menu for the widget.
+    context_menu = Instance("pyface.action.menu_manager.MenuManager")
 
     # ------------------------------------------------------------------------
     # 'IWidget' interface.
@@ -106,3 +109,18 @@ class Widget(MWidget, HasTraits):
     def _set_control_tooltip(self, tooltip):
         """ Toolkit specific method to set the control's tooltip. """
         self.control.SetToolTip(tooltip)
+
+    def _observe_control_context_menu(self, remove=False):
+        """ Toolkit specific method to change the control menu observer. """
+        if remove:
+            self.control.Unbind(
+                wx.EVT_CONTEXT_MENU, handler=self._handle_context_menu
+            )
+        else:
+            self.control.Bind(wx.EVT_CONTEXT_MENU, self._handle_context_menu)
+
+    def _handle_control_context_menu(self, event):
+        """ Signal handler for displaying context menu. """
+        if self.control is not None and self.context_menu is not None:
+            menu = self.context_menu.create_menu(self.control)
+            self.control.PopupMenu(menu)
